Add support of timer tasks for EventLoop on Windows
Most functionalities of TimerQueue, Timer and TimerID can be used on Linux too,
but insead of using timeout of multiplexor, timer_fd would be introduced as the
timer source.

Resolution of TimeDuration varies on Windows and Linux, due to the highest
resolution their timer source can provide. We also change TimePoint definition
to adopt TimeDuration, because resolution of system_clock::time_point is too high,
and there is rounding errors between their conversions to TimeDuration.