Allow naming threads and optimize accesses to thread-id
Use this_thread::SetName() to name a thread; and this setup is also
debugger-friendly in most cases. Thread names are stored in TLS.

Initialization for IOServiceContext now will automatically name main
thread 'ezio-Main'.

Because gettid() on Linux involves a system call, which is not cheap, we
cache the tid of a thread in TLS for performance.