Enable worker-pool support for TCPServer
TCPServer now can enable worker-pool by calling Start() with specific
Options. Each connection is dispatched to a worker-thread and all its IO
operations are performed on that thread.

The selection algorithm currently is simply round-robin, and it is not bad for
load balancing.

TCPServer still runs on main thread.

We add managed notifier bookkeeping on Windows, because though IOCP is
stateful, we can't prevent from queuing completion entries of a dead
socket. So we have to do our own check.

KBase is also upgraded to keep update to date.