Add Connector implementation for Linux
We use connect(2) in non-blocking mode on Linux.

Unlike the implementation on Windows, once a non-fatal error occurred, we
start over from scratch; and self-connected is regarded as a recoverable
failure and simply start Connect() again.

When the Connector enters into wait-for-completion status, i.e. a
notifier is created for waiting IO events, it may end up with EPOLLERROR
and EPOLLOUT both being set, resulting in handlers HandleError() and
HandleNewConnection() both get invoked.